renaming keystonelight to keystone
