+++ .gitignore
@@ -7,3 +7,4 @@ run_tests.log
 covhtml
 pep8.txt
 nosetests.xml
+bla.db
