+++ keystone/backends/sql/__init__.py
@@ -1,0 +1,1 @@
+from keystone.backends.sql.core import *
