+++ keystone/backends/sql/migrate_repo/versions/001_add_initial_tables.py
@@ -1,7 +1,7 @@
 from sqlalchemy import *
 from migrate import *
 
-from keystonelight.backends import sql
+from keystone.backends import sql
 
 
 def upgrade(migrate_engine):
