+++ keystone/catalog.py
@@ -2,8 +2,8 @@
 
 # the catalog interfaces
 
-from keystonelight import config
-from keystonelight import utils
+from keystone import config
+from keystone import utils
 
 
 CONF = config.CONF
