+++ keystone/policy.py
@@ -2,10 +2,8 @@
 
 # the catalog interfaces
 
-import uuid
-
-from keystonelight import config
-from keystonelight import utils
+from keystone import config
+from keystone import utils
 
 
 CONF = config.CONF
