--- keystonelight/backends/sql/__init__.py
@@ -1,1 +1,0 @@
-from keystonelight.backends.sql.core import *
