+++ run_tests.sh
@@ -86,7 +86,7 @@ function run_pep8 {
   ignore_dirs="*ajaxterm*"
   GLOBIGNORE="$ignore_scripts:$ignore_files:$ignore_dirs"
   srcfiles=`find bin -type f ! -name .*.swp`
-  srcfiles+=" keystonelight"
+  srcfiles+=" keystone"
   # Just run PEP8 in current environment
   ${wrapper} pep8 --repeat --show-pep8 --show-source \
     --ignore=E202,E111 \
