+++ tests/backend_sql.conf
@@ -6,4 +6,4 @@ max_pool_size = 10
 pool_timeout = 200
 
 [identity]
-driver = keystonelight.backends.sql.SqlIdentity
+driver = keystone.backends.sql.SqlIdentity
