+++ tests/keystone_compat_diablo.conf
@@ -4,41 +4,41 @@ admin_port = 35357
 admin_token = ADMIN
 
 [identity]
-driver = keystonelight.backends.kvs.KvsIdentity
+driver = keystone.backends.kvs.KvsIdentity
 
 [catalog]
-driver = keystonelight.backends.kvs.KvsCatalog
+driver = keystone.backends.kvs.KvsCatalog
 
 [token]
-driver = keystonelight.backends.kvs.KvsToken
+driver = keystone.backends.kvs.KvsToken
 
 [policy]
-driver = keystonelight.backends.policy.SimpleMatch
+driver = keystone.backends.policy.SimpleMatch
 
 
 [filter:debug]
-paste.filter_factory = keystonelight.wsgi:Debug.factory
+paste.filter_factory = keystone.wsgi:Debug.factory
 
 [filter:token_auth]
-paste.filter_factory = keystonelight.middleware:TokenAuthMiddleware.factory
+paste.filter_factory = keystone.middleware:TokenAuthMiddleware.factory
 
 [filter:admin_token_auth]
-paste.filter_factory = keystonelight.middleware:AdminTokenAuthMiddleware.factory
+paste.filter_factory = keystone.middleware:AdminTokenAuthMiddleware.factory
 
 [filter:json_body]
-paste.filter_factory = keystonelight.middleware:JsonBodyMiddleware.factory
+paste.filter_factory = keystone.middleware:JsonBodyMiddleware.factory
 
-[app:keystonelight]
-paste.app_factory = keystonelight.service:app_factory
+[app:keystone]
+paste.app_factory = keystone.service:app_factory
 
 [app:keystone_service]
-paste.app_factory = keystonelight.keystone_compat:service_app_factory
+paste.app_factory = keystone.keystone_compat:service_app_factory
 
 [app:keystone_admin]
-paste.app_factory = keystonelight.keystone_compat:admin_app_factory
+paste.app_factory = keystone.keystone_compat:admin_app_factory
 
-[pipeline:keystonelight_api]
-pipeline = token_auth admin_token_auth json_body debug keystonelight
+[pipeline:keystone_api]
+pipeline = token_auth admin_token_auth json_body debug keystone
 
 [pipeline:keystone_service_api]
 pipeline = token_auth admin_token_auth json_body debug keystone_service
@@ -48,5 +48,5 @@ pipeline = token_auth admin_token_auth json_body debug keystone_admin
 
 [composite:main]
 use = egg:Paste#urlmap
-/ = keystonelight_api
+/ = keystone_api
 /v2.0 = keystone_service_api
