+++ tests/test_legacy_compat.py
@@ -5,11 +5,11 @@ import sys
 
 from nose import exc
 
-from keystonelight import config
-from keystonelight import logging
-from keystonelight import models
-from keystonelight import test
-from keystonelight import utils
+from keystone import config
+from keystone import logging
+from keystone import models
+from keystone import test
+from keystone import utils
 
 
 CONF = config.CONF
